+++ JinSup/src/Main.java
@@ -5,11 +5,11 @@ public class Main {
   public static void main(String[] args) {
     Scanner scan = new Scanner(System.in);
     System.out.println("Enter the buy price in dollars (increments of 0.25): ");
-    long buyPrice = (long) (scan.nextDouble() * 100);
+    int buyPrice = (int) (scan.nextDouble() * 100);
     System.out.println("Enter the starting period in simulation seconds: ");
-    long startTime = scan.nextLong() * 1000;
+    int startTime = scan.nextInt() * 1000;
     System.out.println("Enter the end time in simulation seconds: ");
-    long endTime = scan.nextLong() * 1000;
+    int endTime = scan.nextInt() * 1000;
     scan.close();
 
     MatchingEngine matchingEngine = new MatchingEngine(buyPrice);
